✨feat(front): improve terminal command execution UX
- add waiting cursor on next line during command execution                                                                        
- hide input form while command is running                                                    
- move `measureTextWidth` to shared `utils.ts`
- fix `scrollToBottom` to use `instant` behavior    
- fix terminal input padding and line-height
- remove trailing space from prompt HTML
